+++ srcs/form_pack2.c
@@ -6,15 +6,15 @@
 /*   By: vquesnel <vquesnel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/08 23:25:23 by vquesnel          #+#    #+#             */
-/*   Updated: 2016/03/22 15:27:39 by vquesnel         ###   ########.fr       */
+/*   Updated: 2016/03/28 20:15:34 by vquesnel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fillit.h"
 
-size_t		ft_is_l(char *map, char c)
+int		ft_is_l(char *map, char c)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (map[i])
@@ -37,9 +37,9 @@ size_t		ft_is_l(char *map, char c)
 	return (0);
 }
 
-size_t		ft_is_j(char *map, char c)
+int		ft_is_j(char *map, char c)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (map[i])
